WIP, web: redesign website layout (esp. front page)
I started playing around with some ideas after work last night, and
figured I should upload my WIP.

The motif is "above the clouds", where the slant is an abstract cloud
layer. The color has been changed to be a not-so-used shade, pulled from
the PatternFly color palette (which we should use across the site,
because: why not?).

I wanted to de-duplicate the big cockpit header, give the (now
shortened) tagline a better place, provide a little bit of fun (with the
airplaines, which act as arrows to the tagline and big screenshot... and happen to imply that you can use it on multiple machines too).

The triangles in the background are an abstract rendering of one of my
photos which has an interesting cloudy sky, ran through G'Mic in GIMP.

Everything below the slant is unfinished. I was experimenting with using
color for the blocks and finding a placement for the thumbnails that
would be more interesting, but stopped there with this (unusable)
result. It will be fixed in the future.

As the top is tightened up with both design and space, we now have a lot
more area to include more compelling features (see issue #146) and
include a bit of info about our audience (issue #143).

Anyway, this is still *very* WIP but I wanted to get it into version
control ASAP.