machines: don't show current snapshot
Spoke to Ran Elimelech and he pointed out that current snapshot
is a weird concept, because it will only be current for a brief
moment when you roll back to it, and then it will diverge for
possibly the end of time. So don't show an indicator for that.

For https://github.com/cockpit-project/cockpit/issues/13010